Tensor that will creted inside convertImageToTensor converted from kFloat32 to kInt8

+++ torch-extension/src/media/opencv/images.cpp
@@ -1,7 +1,6 @@
 #include "../../../include/media/opencv/images.h"
 
 
-
 namespace torch::ext::media::opencv {
     torch::Tensor convertImageToTensor(fs::path img) {
         cv::Mat image = cv::imread(img.string(), cv::IMREAD_COLOR);
@@ -18,8 +17,7 @@ namespace torch::ext::media::opencv {
         // 4. Create a tensor from the image data
         // OpenCV uses HWC (Height, Width, Channels) format
         torch::Tensor tensor = torch::from_blob(image.data, {image.rows, image.cols, image.channels()},
-                                                torch::kFloat32
-        );
+                                                torch::kInt8);
 
         // 5. Permute to CHW (Channels, Height, Width) format, which is PyTorch's default
         tensor = tensor.permute({2, 0, 1});
